move utils tests into tests.py
