--- tests/utils-test.conf
@@ -1,7 +1,0 @@
-#
-# This file is to be used to test
-# mplane.utils module only.
-#
-true_param= True
-false_param= False
-other_param= other
